+++ src/lib/index.ts
@@ -1,3 +1,3 @@
-export { default as ShopkeepThemeProvider } from './ShopkeepThemeProvider.svelte';
-export { default as Button } from './Button.svelte';
-export { default as Icon } from './Icon.svelte';
+export { default as ShopkeepThemeProvider } from './providers/ShopkeepThemeProvider.svelte';
+export { default as Button } from './components/Button.svelte';
+export { default as Icon } from './components/Icon.svelte';
