+++ src/lib/providers/ShopkeepThemeProvider.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import purple from './themes/purple';
+	import purple from '../themes/purple';
 	export let theme: string = 'purple';
 	export let font: string = 'Montserrat';
